refactor: move kafka send msg operation outside provider thread
